+++ packages/models/src/crypto-assets/crypto-asset-info.model.ts
@@ -50,7 +50,6 @@ export interface Src20CryptoAssetInfo extends BaseCryptoAssetInfo {
 // Stacks
 export interface Sip9CryptoAssetInfo {
   readonly contractId: string;
-  readonly contractName: string;
   readonly imageCanonicalUri: string;
   readonly name: string;
 }
@@ -58,7 +57,6 @@ export interface Sip9CryptoAssetInfo {
 export interface Sip10CryptoAssetInfo extends BaseCryptoAssetInfo {
   readonly canTransfer: boolean;
   readonly contractId: string;
-  readonly contractName: string;
   readonly imageCanonicalUri: string;
   readonly name: string;
   readonly symbol: string;