+++ packages/query/src/bitcoin/runes/runes-ticker-info.query.ts
@@ -1,23 +1,29 @@
-import { type UseQueryResult, useQueries } from '@tanstack/react-query';
+import { isDefined } from '@leather-wallet/utils';
+import { useQueries } from '@tanstack/react-query';
 
-import { useConfigRunesEnabled } from '../../common/remote-config/remote-config.query';
 import { useLeatherNetwork } from '../../leather-query-provider';
-import { type RuneTickerInfo, useBitcoinClient } from '../bitcoin-client';
+import { RuneBalance, type RuneTickerInfo, useBitcoinClient } from '../bitcoin-client';
+import { useRunesEnabled } from './runes.hooks';
+import { createRuneCryptoAssetDetails } from './runes.utils';
 
 const queryOptions = { staleTime: 5 * 60 * 1000 };
 
-export function useGetRunesTickerInfoQuery(runeNames: string[]): UseQueryResult<RuneTickerInfo>[] {
+export function useGetRunesTickerInfoQuery(runesBalances: RuneBalance[]) {
   const client = useBitcoinClient();
   const network = useLeatherNetwork();
-  const runesEnabled = useConfigRunesEnabled();
+  const runesEnabled = useRunesEnabled();
 
   return useQueries({
-    queries: runeNames.map(runeName => {
+    queries: runesBalances.map(runeBalance => {
       return {
-        enabled: !!runeName && (network.chain.bitcoin.bitcoinNetwork === 'testnet' || runesEnabled),
-        queryKey: ['runes-ticker-info', runeName],
+        enabled: isDefined(runeBalance) && runesEnabled,
+        queryKey: ['runes-ticker-info', runeBalance.rune_name],
         queryFn: () =>
-          client.BestinSlotApi.getRunesTickerInfo(runeName, network.chain.bitcoin.bitcoinNetwork),
+          client.BestinSlotApi.getRunesTickerInfo(
+            runeBalance.rune_name,
+            network.chain.bitcoin.bitcoinNetwork
+          ),
+        select: (resp: RuneTickerInfo) => createRuneCryptoAssetDetails(runeBalance, resp),
         ...queryOptions,
       };
     }),